Revert "Merge branch 'master' into master"
This reverts commit 476b50671803ad9af10b311252da63b92334a115, reversing
changes made to 3e6139efebf1c90ba763720ef2fcb3128d8dc59b.

+++ Assets/scripts/EntityClasses/FuelDispenser.cs
@@ -7,7 +7,7 @@ namespace EntityClasses
 {
     public class FuelDispenser : EntityInterface
     {
-        int dispensingSpeed;
+        double dispensingSpeed;
 
         public void Add(List<EntityInterface> list)
         {
